--- com.unity.renderstreaming/Runtime/Scripts/AudioSourceStreamSender.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 28e4557ba7cb056499034647b21b6361
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
